refactor: categorize components into some directories
