--- src/components/index.css
@@ -1,70 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@100;400;700&display=swap');
-
-html {
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  font-smoothing: antialiased;
-}
-
-/* Reset */
-*, *::before, *::after { box-sizing: border-box; }
-article > * + * { margin-top: 1em; }
-input, button, textarea, select { font: inherit; }
-
-body, h1, h2, h3, h4, p,
-ul[class], ol[class], li,
-figure, figcaption, blockquote,
-dl, dd { margin: 0; }
-
-ul[class], ol[class] {
-  padding: 0;
-  list-style: none;
-}
-
-body {
-  min-height: 100vh;
-  scroll-behavior: smooth;
-  text-rendering: optimizeSpeed;
-  line-height: 1.5;
-
-  font-family: 'Roboto', sans-serif;
-  font-style: normal;
-  font-weight: 400;
-}
-
-img {
-  max-width: 100%;
-  display: block;
-}
-
-@media (prefers-reduced-motion: reduce) {
-  * {
-    animation-duration: 0.01ms !important;
-    animation-iteration-count: 1 !important;
-    transition-duration: 0.01ms !important;
-    scroll-behavior: auto !important;
-  }
-}
-
-/* Global-ish */
-h1, h2 {
-  font-family: 'Roboto', sans-serif;
-  font-weight: bold;
-}
-
-h1 {
-  font-size: 64px;
-  line-height: 75px;
-  margin-bottom: 45px;
-}
-
-h2 {
-  font-size: 28px;
-  line-height: 33px;
-}
-
-p {
-  font-size: 28px;
-  line-height: 33px;
-  margin-bottom: 45px;
-}
